feat: Add pagination support to API list endpoints

+++ cli/internal/api/client.go
@@ -42,10 +42,10 @@ type Response[T any] struct {
 }
 
 type Pagination struct {
-	Page       int `json:"page"`
-	Limit      int `json:"limit"`
-	Total      int `json:"total"`
-	TotalPages int `json:"totalPages"`
+	Page       int   `json:"page"`
+	Limit      int   `json:"limit"`
+	Total      int64 `json:"total"`
+	TotalPages int   `json:"totalPages"`
 }
 
 // APIError is returned when the server sends a non-2xx status.
